Data type change does not indicate that there has been a change Changing the order of arguments in an activity diagram does not count as a change Improved package placement Documentation improvements
