+++ doc/DebugConfiguration.md
@@ -1,8 +1,8 @@
-# summary
+# Summary
 
 For debugging purposes, this document describes how to change the logging settings
 
-## Configure RPExtensionUtilities.hep
+# Configure RPExtensionUtilities.hep
 
 Add one of the following to the `name` of each plugin's configuration in RPExtensionUtilities.hep.  
 * `LogDebug`:Outputs debug-level logs when executing commands.
