+++ doc/RPDoxygenXML.md
@@ -1,10 +1,43 @@
 # Plug-in Features
 
-Import from Doxygen's XML output from C source into Rhapsody's model.  
-Create packages for each version to manage API changes and deletions.  
+## Summary
+Generate method activity diagrams and callback events using as input an XML file generated by doxygen based on C source code.
+
 Detects activity diagrams that use modified or deleted APIs.  
 
-## Configure RPExtensionUtilities.hep
+
+## Structure
+The structure to be generated is as follows
+```
+[target-package] Right click to select
+   [version name] from user specified
+      Module (more than 1) from doxygen
+         * Activity diagram created from doxygen methods 
+         * Events created from doxygen callbacks
+         * DataType and ValueType created from doxygen types
+```
+Allow users to specify version when importing XML files.  
+The imported and generated APIs and data types should be placed under the latest imported version of the package.
+
+## Changes and Deletions
+If a method or callback argument is changed, the element with the older version is renamed as follows
+```
+   changed_[version_name]_API_name
+   deleted_[version name]_API name
+```
+If an element with the same name already exists in the same version,  
+the name is not changed and the change is reflected.
+
+## Tags to be assigned to the element
+Added/Modified/Deleted API and data types should be given the following tags.  
+* `FirstApplicableVersion`: Set the first imported version.
+* `UnavailableStartVersion`: Set the version that was changed or deleted.
+* `Excluded`: Set the target to be excluded from modification or deletion on import.   
+  **Note**
+  This tag must be assigned by the user.
+
+
+# Configure RPExtensionUtilities.hep
 
 Add the following to RPExtensionUtilities.hep.  
 Modify the 1 in each item to match the number of plug-ins/helpers described in the file.  
@@ -30,9 +63,9 @@ isVisible2=1
 DLLServerCompatible2=1
 ```
 
-## Rhapsody operations
+# Rhapsody operations
 
-### Import from Doxygen XML to Rhapsody
+## Import from Doxygen XML to Rhapsody
 
 1. Select the package of IBM Rhapsody explorer.
 2. Right-click to display menu.
@@ -46,7 +79,7 @@ DLLServerCompatible2=1
 8. The import status will be displayed in the output window.
 9. A popup notifies you when the import is finished.
 
-### Check Unavailable Activities
+## Check Unavailable Activities
 
 1. Select the package of IBM Rhapsody explorer.
 2. Right-click to display menu.
