+++ doc/RPImageOutActivity.md
@@ -3,7 +3,7 @@
 Outputs an activity diagram under the specified package as an image file.  
 The output destination of the image file is the same directory as the project file of the active project.  
 
-## Configure RPExtensionUtilities.hep
+# Configure RPExtensionUtilities.hep
 
 Add the following to RPExtensionUtilities.hep.  
 Modify the 1 in each item to match the number of plug-ins/helpers described in the file.  
@@ -20,7 +20,7 @@ isVisible1=1
 DLLServerCompatible1=1
 ```
 
-## Rhapsody operations
+# Rhapsody operations
 
 1. Select the package of IBM Rhapsody explorer.
 2. Right-click to display menu.
