+++ src/com/ibm/rhapsody/rputilities/rpcommand/importer/bridge/RPBridgeEnum.java
@@ -3,10 +3,10 @@ package com.ibm.rhapsody.rputilities.rpcommand.importer.bridge;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.ibm.rhapsody.rputilities.doxygen.type.DoxygenType;
-import com.ibm.rhapsody.rputilities.doxygen.TAGTYPE;
-import com.ibm.rhapsody.rputilities.doxygen.type.DoxygenTypeEnumValue;
 import com.ibm.rhapsody.rputilities.rpcommand.importer.RPTYPE_KIND;
+import com.ibm.rhapsody.rputilities.rpcommand.importer.doxygen.TAGTYPE;
+import com.ibm.rhapsody.rputilities.rpcommand.importer.doxygen.type.DoxygenType;
+import com.ibm.rhapsody.rputilities.rpcommand.importer.doxygen.type.DoxygenTypeEnumValue;
 import com.telelogic.rhapsody.core.IRPEnumerationLiteral;
 import com.telelogic.rhapsody.core.IRPModelElement;
 import com.telelogic.rhapsody.core.IRPPackage;
@@ -90,10 +90,10 @@ public class RPBridgeEnum extends ARPBridge {
             return true;
         }
 
-        if (checkUpdate(GetKind(), rpType.getKind()) == true) {
-            trace(getName() + " change Kind " + rpType.getKind() + "->" + GetKind());
-            return true;
-        }
+        // if (checkUpdate(GetKind(), rpType.getKind()) == true) {
+        //     trace(getName() + " change Kind " + rpType.getKind() + "->" + GetKind());
+        //     return true;
+        // }
 
         return false;
     }
@@ -111,17 +111,17 @@ public class RPBridgeEnum extends ARPBridge {
         IRPType rpType = getObject(element);
 
         if (checkUpdate(getName(), rpType.getDisplayName()) == true) {
-            trace(getName() + " apply DisplayName " + rpType.getDisplayName() + "->" + getName());
+            debug(getName() + " apply DisplayName " + rpType.getDisplayName() + "->" + getName());
             rpType.setDisplayName(getName());
         }
 
         if (checkUpdate(getName(), rpType.getName()) == true) {
-            trace(getName() + " apply Name " + rpType.getName() + "->" + getName());
+            debug(getName() + " apply Name " + rpType.getName() + "->" + getName());
             rpType.setName(getName());
         }
 
         if (checkUpdate(GetKind(), rpType.getKind()) == true) {
-            trace(getName() + " apply Kind " + rpType.getKind() + "->" + GetKind());
+            debug(getName() + " apply Kind " + rpType.getKind() + "->" + GetKind());
             rpType.setKind(GetKind());
         }
 
@@ -139,10 +139,17 @@ public class RPBridgeEnum extends ARPBridge {
     }
 
     protected void applyEnumValue(IRPType rpType, DoxygenType value, String currentVersion) {
-        DoxygenTypeEnumValue enumvalue = getObject(value);
-        // info(rpType.getKind() + ":" + GetKind()+ " applyEnumValue:"+value.getName());
-        IRPEnumerationLiteral literal = rpType.addEnumerationLiteral(value.getName());
-        literal.setValue(enumvalue.getInitializer());
+        
+        try {
+            DoxygenTypeEnumValue enumvalue = getObject(value);
+            debug(rpType.getName() + " applyEnumValue:"+value.getName());
+            IRPEnumerationLiteral literal = rpType.addEnumerationLiteral(value.getName());
+            literal.setValue(enumvalue.getInitializer());
+        } catch (Exception e) {
+            error("applyEnumValue name:" + rpType.getName() + " applyEnumValue:"+ value.getName(), e);
+            doxygen_.logoutdebug(0);
+        }
+
         return;
     }
