+++ src/com/ibm/rhapsody/rputilities/rpcommand/importer/doxygen/DoxygenXMLParseOption.java
@@ -1,11 +1,11 @@
-package com.ibm.rhapsody.rputilities.doxygen;
+package com.ibm.rhapsody.rputilities.rpcommand.importer.doxygen;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import javax.xml.stream.XMLStreamReader;
 
-import com.ibm.rhapsody.rputilities.doxygen.type.DoxygenType;
+import com.ibm.rhapsody.rputilities.rpcommand.importer.doxygen.type.DoxygenType;
 import com.ibm.rhapsody.rputilities.rpcore.ARPObject;
 
 public class DoxygenXMLParseOption extends ARPObject {
